requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-language==1.3.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
imutils==0.5.3
MouseInfo==0.0.4
nltk==3.4.5
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
PyAutoGUI==0.9.47
PyGetWindow==0.0.7
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.5
